Fix #34 - Introduced get booleanAttributes() { return ['name']; }
It looks like last change created a regression for the most common
non boolean attribute case.

Since it's not possible to understand upfront the nature of an attribute,
the proposed implementation of boolean attributes is based on a getter
defined same way observedAttributes is defined, containing a list
of attributes that should be threated as booleans.

If the set value is falsy or the exact string "false" the attreibute
is removed from the element, otherwise is set with whatever passed value.