Fix built-in elements inheritance
Due to built-in elements extending the standard HTML Classes,
there's a need to add HyperHTML-Element's own class members as well.

HyperHTMLElement assumed the Super class was going to be itself,
this commit fixes that by explicitly adding HyperHTMLElement
own keys to the Intermediate class used to create our own built-in
element extension.