Do not override the getter of a boolean attribute that is written in kebab case.
The problem is that the define method checks the attribute
in the prototype using its name in the kebab case,
but defines a getter in the camel case.
This leads to the problem that a defined getter for a boolean attribute
can be overridden by a normal getter for an observed attribute.