Update TypeScript definition of 'get defaultState'.
Now it's impossible to define getter for 'defaultState' with TypeScript.
Following error appears:
"'defaultState' is defined as a property in class 'HyperHTMLElement<{}>', but is overridden here in 'MyComponent' as an accessor."
See full description in issue: https://github.com/WebReflection/hyperHTML-Element/issues/77

This is because defaultState is defined as just readonly property
when it should be described as getter.
Description for 'defaultState' was updated to define it as 'getter'